move few forum-related files in the forum features
