+++ src/app/forum/containers/thread-container/thread-container.component.scss
