+++ src/app/forum/containers/thread-message/thread-message.component.scss
