+++ src/app/forum/containers/thread/thread.component.html
